sam-agents/pyproject.toml
[project]
name = "flowback-sam-agents"
version = "0.1.0"
description = "Solace Agent Mesh agents for Flowback analytics"
requires-python = ">=3.11"
dependencies = [
    "solace-agent-mesh==1.13.3",
    "nats-py==2.8.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[tool.uv]
python-version = "3.11"
